LayoutTests/http/tests/webgpu/webgpu/api/operation/rendering/robust_access_index.html


                <script>
                    if (window.testRunner)
                        testRunner.waitUntilDone();
                    window.location.href = "/webgpu/cts.https.html?q=webgpu:api,operation,rendering,robust_access_index:*";
                </script>
            
